automation/requirements.txt
# Core testing frameworks
pytest>=7.4.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
pytest-mock>=3.11.1

# HTTP and API testing
requests>=2.31.0
aiohttp>=3.8.5
httpx>=0.24.1

# Database connectivity
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.20

# Image processing and computer vision
opencv-python>=4.8.0.76
Pillow>=10.0.0
numpy>=1.24.3

# Screenshot comparison and validation
scikit-image>=0.21.0
matplotlib>=3.7.2

# Report generation
jinja2>=3.1.2
markdown>=3.4.4
weasyprint>=59.0

# Automation and browser control
selenium>=4.11.2
webdriver-manager>=4.0.0

# Data processing and analysis
pandas>=2.0.3
openpyxl>=3.1.2

# Configuration and environment
python-decouple>=3.8
pyyaml>=6.0.1
toml>=0.10.2

# Logging and monitoring
structlog>=23.1.0
colorlog>=6.7.0

# Async support
asyncio>=3.4.3
asyncpg>=0.28.0

# Development and debugging
ipython>=8.14.0
pdbpp>=0.10.3

# Performance monitoring
memory-profiler>=0.61.0
psutil>=5.9.5

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0

# Security
bandit>=1.7.5
safety>=2.3.5

# Utilities
click>=8.1.7
rich>=13.5.2
tqdm>=4.66.1
